refactor: get rid of MutableEntity
